TEIID-360 TEIID-359 TEIID-357 TEIID-356 TEIID-354 TEIID-353 TEIID-352 TEIID-164 jdbc connector api cleanup.  created canonical push-down function list, changed the SQLTranslator to handle the tasks that were previously in specific SQLConversionVisitors.

+++ connectors/connector-jdbc/src/main/java/com/metamatrix/connector/jdbc/sybase/SybaseConvertModifier.java
@@ -408,7 +408,7 @@ public class SybaseConvertModifier extends BasicFunctionModifier implements Func
                     new IExpression[] { 
                         langFactory.createLiteral("varchar", String.class),  //$NON-NLS-1$
                         args[0],
-                        langFactory.createLiteral(new Integer(1), Integer.class) },
+                        langFactory.createLiteral(new Integer(109), Integer.class) },
                     String.class);
                 
                 // Build outer convert
